#4 [MOVE] : example -> example.xml 위치 이동
