+++ app/src/main/java/com/univoice/feature/example/xml/ExampleActivity.kt
@@ -1,4 +1,4 @@
-package com.univoice.feature.example
+package com.univoice.feature.example.xml
 
 import androidx.activity.viewModels
 import androidx.lifecycle.flowWithLifecycle
